Added tyre information display and TTS capabilities, rearchitecture of application (#33)
Race screen now shows tyre history. Qt TTS was integrated to inform the user of tyre changes, featuring calls by driver name, if available, and by position if not. This functionality was extended to penalty information and fastest lap as well.

Major rearchitecture of application also done; internal packets were separated into internal to event. A new layer called generalizer was introduced between NetCom and Processor to convert game packets into internal packets; the processor takes the internal packets still and dissects the information to generate event packets that are then sent to the UI. This is more modular, more easily understandable, easier to manage the memory of, and avoids packet confusion.